- Report all outliers of all stations in file - Separate analysis_interval functions

+++ jaziku/modules/climate/lags.py
@@ -24,6 +24,7 @@ from calendar import monthrange
 from datetime import date
 from dateutil.relativedelta import relativedelta
 
+from jaziku.modules.analysis_interval import get_values_in_range_analysis_interval, get_range_analysis_interval
 from jaziku.utils import globals_vars
 from jaziku.utils import format_out
 from jaziku.utils.mean import mean
@@ -48,107 +49,14 @@ def get_lag_values(station, var, lag, month, day=None):
     return [row[var_select[var]] for row in temp_list]
 
 
-def get_range_analysis_interval(station):
-    # range based on analysis interval
-
-    if globals_vars.config_run['analysis_interval'] != "trimester":
-        if globals_vars.analysis_interval_num_days == 5:
-            return [1, 6, 11, 16, 21, 26]
-        if globals_vars.analysis_interval_num_days == 10:
-            return [1, 11, 21]
-        if globals_vars.analysis_interval_num_days == 15:
-            return [1, 16]
-    else:
-        return None
-
-def get_range_analysis_interval_values(station, type, year, month, day=None, lag=None):
-    """
-    Get all values inside range analysis interval in specific year, month, day or lag.
-    The "type" must be "D" or "I". For trimester the "month" is the start month of
-    trimester, for 5, 10, or 15 days the "day" is a start day in the range analysis
-    interval. The "lag" is only affect for the independent variable and it must
-    be 0, 1 or 2.
-    """
-
-    range_analysis_interval = get_range_analysis_interval(station)
-
-    if type == 'D':
-        var_D_values = []
-        if station.var_D.frequency_data == "daily":
-            # clone range for add the last day (32) for calculate interval_day_var_D
-            rai_plus = list(range_analysis_interval)
-            rai_plus.append(32)
-            # from day to next iterator based on analysis interval
-            # e.g. [0,1,2,3,4] for first iteration for 5 days
-            interval_day_var_D = range(day - 1, rai_plus[rai_plus.index(day) + 1] - 1)
-
-            for iter_day in interval_day_var_D:
-                now = date(year, month, 1) + relativedelta(days=iter_day)
-                # check if continues with the same month
-                if now.month == month:
-                    index_var_D = station.var_D.date.index(now)
-                    var_D_values.append(station.var_D.data[index_var_D])
-        if station.var_D.frequency_data == "monthly":
-            # get the three values for var_D in this month
-            for iter_month in range(3):
-                var_D_values.append(station.var_D.data[station.var_D.date.index(
-                    date(year, month, 1) + relativedelta(months=iter_month))])
-
-        return var_D_values
-
-    if type == 'I':
-        var_I_values = []
-        if station.var_I.frequency_data == "daily":
-            # from day to next iterator based on analysis interval
-            start_interval = range_analysis_interval[range_analysis_interval.index(day) - lag]
-            try:
-                end_interval = range_analysis_interval[range_analysis_interval.index(day) + 1 - lag]
-            except:
-                end_interval = range_analysis_interval[0]
-
-            start_date = date(year, month, start_interval)
-
-            if range_analysis_interval.index(day) - lag < 0:
-                start_date += relativedelta(months= -1)
-
-            iter_date = start_date
-
-            while iter_date.day != end_interval:
-                index_var_I = station.var_I.date.index(iter_date)
-                var_I_values.append(station.var_I.data[index_var_I])
-                iter_date += relativedelta(days=1)
-
-        if station.var_I.frequency_data == "monthly":
-            if station.state_of_data in [1, 3]:
-                # get the three values for var_I in this month
-                for iter_month in range(3):
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(year, month, 1) + relativedelta(months=iter_month - lag))])
-            if station.state_of_data in [2]:
-                # keep constant value for month
-                if globals_vars.config_run['analysis_interval'] == "trimester":
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(year, month, 1) + relativedelta(months= -lag))])
-                else:
-                    real_date = date(year, month, day) + relativedelta(days= -globals_vars.analysis_interval_num_days * lag)
-                    # e.g if lag 2 in march and calculate to 15days go to february and not january
-                    if month - real_date.month > 1:
-                        real_date = date(real_date.year, real_date.month + 1, 1)
-
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(real_date.year, real_date.month, 1))])
-
-        return var_I_values
-
-
-def calculate_lags(station):
+def calculate_lags(station, makes_files=True):
     """
     Calculate and return lags 0, 1 and 2 of specific stations
     and save csv file of time series for each lag and trimester,
     the lags are calculated based in: +1 year below of start
     common period and -1 year above of end common period.
 
-    Return:
+    Return by reference:
     station with: Lag_0, Lag_1, Lag_2 and range_analysis_interval
     """
 
@@ -158,10 +66,11 @@ def calculate_lags(station):
     Lag_1 = []
     Lag_2 = []
 
-    # directories to save lags
-    dir_lag = [os.path.join(station.climate_dir, _('time_series'), _('lag_0')),
-               os.path.join(station.climate_dir, _('time_series'), _('lag_1')),
-               os.path.join(station.climate_dir, _('time_series'), _('lag_2'))]
+    if makes_files:
+        # directories to save lags
+        dir_lag = [os.path.join(station.climate_dir, _('time_series'), _('lag_0')),
+                   os.path.join(station.climate_dir, _('time_series'), _('lag_1')),
+                   os.path.join(station.climate_dir, _('time_series'), _('lag_2'))]
 
     range_analysis_interval = get_range_analysis_interval(station)
 
@@ -169,27 +78,29 @@ def calculate_lags(station):
 
         for lag in globals_vars.lags:
 
-            if not os.path.isdir(dir_lag[lag]):
-                os.makedirs(dir_lag[lag])
+            if makes_files:
+                if not os.path.isdir(dir_lag[lag]):
+                    os.makedirs(dir_lag[lag])
 
             # all months in year 1->12
             for month in range(1, 13):
-                csv_name = os.path.join(dir_lag[lag],
-                    _('Mean_lag_{0}_trim_{1}_{2}_{3}_{4}_{5}_'
-                      '({6}-{7}).csv')
-                    .format(lag, month, station.code,
-                        station.name, station.type_D,
-                        station.type_I,
-                        station.process_period['start'],
-                        station.process_period['end']))
+                if makes_files:
+                    csv_name = os.path.join(dir_lag[lag],
+                        _('Mean_lag_{0}_trim_{1}_{2}_{3}_{4}_{5}_'
+                          '({6}-{7}).csv')
+                        .format(lag, month, station.code,
+                            station.name, station.type_D,
+                            station.type_I,
+                            station.process_period['start'],
+                            station.process_period['end']))
 
-                if os.path.isfile(csv_name):
-                    os.remove(csv_name)
+                    if os.path.isfile(csv_name):
+                        os.remove(csv_name)
 
-                # output write file:
-                # [[ yyyy/month, Mean_Lag_X_var_D, Mean_Lag_X_var_I ],... ]
-                open_file = open(csv_name, 'w')
-                csv_file = csv.writer(open_file, delimiter=';')
+                    # output write file:
+                    # [[ yyyy/month, Mean_Lag_X_var_D, Mean_Lag_X_var_I ],... ]
+                    open_file = open(csv_name, 'w')
+                    csv_file = csv.writer(open_file, delimiter=';')
 
                 iter_year = station.process_period['start']
 
@@ -198,10 +109,10 @@ def calculate_lags(station):
                 while iter_year <= station.process_period['end']:
 
                     # get values and calculate mean_var_D
-                    mean_var_D = mean(get_range_analysis_interval_values(station,'D', iter_year, month))
+                    mean_var_D = mean(get_values_in_range_analysis_interval(station,'D', iter_year, month))
 
                     # get values and calculate mean_var_I
-                    mean_var_I = mean(get_range_analysis_interval_values(station,'I', iter_year, month, lag=lag))
+                    mean_var_I = mean(get_values_in_range_analysis_interval(station,'I', iter_year, month, lag=lag))
 
                     # add line in list: Lag_X
                     vars()['Lag_' + str(lag)].append([date(iter_year, month, 1),
@@ -209,41 +120,44 @@ def calculate_lags(station):
                                                       mean_var_I])
 
                     # add line output file csv_file
-                    csv_file.writerow([str(iter_year) + "/" + str(month),
-                                       format_out.number(mean_var_D),
-                                       format_out.number(mean_var_I)])
+                    if makes_files:
+                        csv_file.writerow([str(iter_year) + "/" + str(month),
+                                           format_out.number(mean_var_D),
+                                           format_out.number(mean_var_I)])
                     # next year
                     iter_year += 1
 
-                open_file.close()
-                del csv_file
+                if makes_files:
+                    open_file.close()
+                    del csv_file
 
     if station.state_of_data in [2, 4]:
 
         for lag in globals_vars.lags:
-
-            if not os.path.isdir(dir_lag[lag]):
-                os.makedirs(dir_lag[lag])
+            if makes_files:
+                if not os.path.isdir(dir_lag[lag]):
+                    os.makedirs(dir_lag[lag])
 
             # all months in year 1->12
             for month in range(1, 13):
-                csv_name = os.path.join(dir_lag[lag],
-                    _('Mean_lag_{0}_{1}days_month_{2}_{3}_'
-                      '{4}_{5}_{6}_({7}-{8}).csv')
-                    .format(lag, globals_vars.analysis_interval_num_days,
-                        month, station.code,
-                        station.name, station.type_D,
-                        station.type_I,
-                        station.process_period['start'],
-                        station.process_period['end']))
+                if makes_files:
+                    csv_name = os.path.join(dir_lag[lag],
+                        _('Mean_lag_{0}_{1}days_month_{2}_{3}_'
+                          '{4}_{5}_{6}_({7}-{8}).csv')
+                        .format(lag, globals_vars.analysis_interval_num_days,
+                            month, station.code,
+                            station.name, station.type_D,
+                            station.type_I,
+                            station.process_period['start'],
+                            station.process_period['end']))
 
-                if os.path.isfile(csv_name):
-                    os.remove(csv_name)
+                    if os.path.isfile(csv_name):
+                        os.remove(csv_name)
 
-                # output write file:
-                # [[ yyyy/month, Mean_Lag_X_var_D, Mean_Lag_X_var_I ],... ]
-                open_file = open(csv_name, 'w')
-                csv_file = csv.writer(open_file, delimiter=';')
+                    # output write file:
+                    # [[ yyyy/month, Mean_Lag_X_var_D, Mean_Lag_X_var_I ],... ]
+                    open_file = open(csv_name, 'w')
+                    csv_file = csv.writer(open_file, delimiter=';')
 
                 #days_for_this_month = monthrange(iter_year, month)[1]
 
@@ -261,10 +175,10 @@ def calculate_lags(station):
                             continue
 
                         # get values and calculate mean_var_D
-                        mean_var_D = mean(get_range_analysis_interval_values(station,'D', iter_year, month, day, lag))
+                        mean_var_D = mean(get_values_in_range_analysis_interval(station,'D', iter_year, month, day, lag))
 
                         # get values and calculate mean_var_I
-                        mean_var_I = mean(get_range_analysis_interval_values(station,'I', iter_year, month, day, lag))
+                        mean_var_I = mean(get_values_in_range_analysis_interval(station,'I', iter_year, month, day, lag))
 
                         # add line in list: Lag_X
                         vars()['Lag_' + str(lag)].append([date(iter_year, month, day),
@@ -272,15 +186,16 @@ def calculate_lags(station):
                                                           mean_var_I])
 
                         # add line output file csv_file
-                        csv_file.writerow([str(iter_year) + "/" + str(month)
-                                           + "/" + str(day),
-                                           format_out.number(mean_var_D),
-                                           format_out.number(mean_var_I)])
+                        if makes_files:
+                            csv_file.writerow([str(iter_year) + "/" + str(month)
+                                               + "/" + str(day),
+                                               format_out.number(mean_var_D),
+                                               format_out.number(mean_var_I)])
                         # next year
                         iter_year += 1
-
-                open_file.close()
-                del csv_file
+                if makes_files:
+                    open_file.close()
+                    del csv_file
 
     station.range_analysis_interval = range_analysis_interval
     station.Lag_0, station.Lag_1, station.Lag_2 = Lag_0, Lag_1, Lag_2
